# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D-WISE Tool Suite Backend API",
    "description": "The REST API for the D-WISE Tool Suite Backend.",
    "version": "beta_mwp_1"
  },
  "paths": {
...
    "/search/sdoc_info": {
      "post": {
        "tags": ["search"],
        "summary": "Returns Search Info.",
        "operationId": "search_sdocs_info",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "title": "Project Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ColumnInfo_SearchColumns_" },
                  "title": "Response Search-Search Sdocs Info"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/search/sdoc": {
      "post": {
        "tags": ["search"],
        "summary": "Returns all SourceDocument Ids that match the query parameters.",
        "operationId": "search_sdocs",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "search_query",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Search Query" }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "title": "Project Id" }
          },
          {
            "name": "expert_mode",
            "in": "query",
            "required": true,
            "schema": { "type": "boolean", "title": "Expert Mode" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Body_search-search_sdocs" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "type": "integer" }, "title": "Response Search-Search Sdocs" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/search/code_stats": {
      "post": {
        "tags": ["search"],
        "summary": "Returns SpanEntityStats for the given SourceDocuments.",
        "operationId": "search_code_stats",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "code_id", "in": "query", "required": true, "schema": { "type": "integer", "title": "Code Id" } },
          {
            "name": "sort_by_global",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false, "title": "Sort By Global" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "array", "items": { "type": "integer" }, "title": "Sdoc Ids" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SpanEntityStat" },
                  "title": "Response Search-Search Code Stats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/search/keyword_stats": {
      "post": {
        "tags": ["search"],
        "summary": "Returns KeywordStats for the given SourceDocuments.",
        "operationId": "search_keyword_stats",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "title": "Project Id" }
          },
          {
            "name": "sort_by_global",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false, "title": "Sort By Global" }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50, "title": "Top K" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "array", "items": { "type": "integer" }, "title": "Sdoc Ids" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/KeywordStat" },
                  "title": "Response Search-Search Keyword Stats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/search/tag_stats": {
      "post": {
        "tags": ["search"],
        "summary": "Returns Stat for the given SourceDocuments.",
        "operationId": "search_tag_stats",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "sort_by_global",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false, "title": "Sort By Global" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "array", "items": { "type": "integer" }, "title": "Sdoc Ids" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TagStat" },
                  "title": "Response Search-Search Tag Stats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/search/lexical/memo/content": {
      "post": {
        "tags": ["search"],
        "summary": "Returns all Memos where the content matches the query via lexical search",
        "operationId": "search_memos_by_content_query",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer", "maximum": 10000000, "minimum": 0 }, { "type": "null" }],
              "title": "Skip",
              "description": "The number of elements to skip (offset)"
            },
            "description": "The number of elements to skip (offset)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer", "maximum": 1000, "minimum": 1 }, { "type": "null" }],
              "title": "Limit",
              "description": "The maximum number of returned elements"
            },
            "description": "The maximum number of returned elements"
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoContentQuery" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedMemoSearchResults" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/search/simsearch/sentences": {
      "post": {
        "tags": ["search"],
        "summary": "Returns similar sentences according to a textual or visual query.",
        "operationId": "find_similar_sentences",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SimSearchQuery" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/SimSearchSentenceHit" },
                  "type": "array",
                  "title": "Response Search-Find Similar Sentences"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/search/simsearch/images": {
      "post": {
        "tags": ["search"],
        "summary": "Returns similar images according to a textual or visual query.",
        "operationId": "find_similar_images",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SimSearchQuery" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/SimSearchImageHit" },
                  "type": "array",
                  "title": "Response Search-Find Similar Images"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachedObjectType": {
        "type": "string",
        "enum": [
          "annotation_document",
          "source_document",
          "code",
          "span_annotation",
          "span_group",
          "bbox_annotation",
          "project",
          "document_tag"
        ],
        "title": "AttachedObjectType"
      },
...
      "Body_search-search_sdocs": {
        "properties": {
          "filter": { "$ref": "#/components/schemas/Filter_SearchColumns_" },
          "sorts": {
            "items": { "$ref": "#/components/schemas/Sort_SearchColumns_" },
            "type": "array",
            "title": "Sorts"
          }
        },
        "type": "object",
        "required": ["filter", "sorts"],
        "title": "Body_search-search_sdocs"
      },
      "BooleanOperator": {
        "type": "string",
        "enum": ["BOOLEAN_EQUALS", "BOOLEAN_NOT_EQUALS"],
        "title": "BooleanOperator"
      },
...
      "ColumnInfo_SearchColumns_": {
        "properties": {
          "label": { "type": "string", "title": "Label" },
          "column": {
            "anyOf": [{ "$ref": "#/components/schemas/SearchColumns" }, { "type": "integer" }],
            "title": "Column"
          },
          "sortable": { "type": "boolean", "title": "Sortable" },
          "operator": { "$ref": "#/components/schemas/FilterOperator" },
          "value": { "$ref": "#/components/schemas/FilterValueType" }
        },
        "type": "object",
        "required": ["label", "column", "sortable", "operator", "value"],
        "title": "ColumnInfo[SearchColumns]"
      },
...
      "DateOperator": {
        "type": "string",
        "enum": ["DATE_EQUALS", "DATE_GT", "DATE_LT", "DATE_GTE", "DATE_LTE"],
        "title": "DateOperator"
      },
...
      "DocumentTagRead": {
        "properties": {
          "title": { "type": "string", "title": "Title", "description": "Title of the DocumentTag" },
          "color": { "type": "string", "title": "Color", "description": "Color of the DocumentTag" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description",
            "description": "Description of the DocumentTag"
          },
          "parent_tag_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Parent Tag Id",
            "description": "Parent of the DocumentTag"
          },
          "id": { "type": "integer", "title": "Id", "description": "ID of the DocumentTag" },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "Project the DocumentTag belongs to"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created",
            "description": "Created timestamp of the DocumentTag"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "title": "Updated",
            "description": "Updated timestamp of the DocumentTag"
          }
        },
        "type": "object",
        "required": ["title", "color", "id", "project_id", "created", "updated"],
        "title": "DocumentTagRead"
      },
...
      "FilterExpression_SearchColumns_": {
        "properties": {
          "column": {
            "anyOf": [{ "$ref": "#/components/schemas/SearchColumns" }, { "type": "integer" }],
            "title": "Column"
          },
          "operator": {
            "anyOf": [
              { "$ref": "#/components/schemas/IDOperator" },
              { "$ref": "#/components/schemas/NumberOperator" },
              { "$ref": "#/components/schemas/StringOperator" },
              { "$ref": "#/components/schemas/IDListOperator" },
              { "$ref": "#/components/schemas/ListOperator" },
              { "$ref": "#/components/schemas/DateOperator" },
              { "$ref": "#/components/schemas/BooleanOperator" }
            ],
            "title": "Operator"
          },
          "value": {
            "anyOf": [
              { "type": "boolean" },
              { "type": "string" },
              { "type": "integer" },
              { "items": { "type": "string" }, "type": "array" },
              { "items": { "items": { "type": "string" }, "type": "array" }, "type": "array" }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": ["column", "operator", "value"],
        "title": "FilterExpression[SearchColumns]"
      },
...
      "FilterOperator": {
        "type": "string",
        "enum": ["BOOLEAN", "STRING", "ID", "NUMBER", "ID_LIST", "LIST", "DATE"],
        "title": "FilterOperator"
      },
      "FilterValueType": {
        "type": "string",
        "enum": ["SDOC_ID", "CODE_ID", "USER_ID", "TAG_ID", "SPAN_ANNOTATION", "DOC_TYPE", "INFER_FROM_OPERATOR"],
        "title": "FilterValueType"
      },
...
      "Filter_SearchColumns_": {
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                { "$ref": "#/components/schemas/FilterExpression_SearchColumns_" },
                { "$ref": "#/components/schemas/Filter_SearchColumns_" }
              ]
            },
            "type": "array",
            "title": "Items"
          },
          "logic_operator": { "$ref": "#/components/schemas/LogicalOperator" }
        },
        "type": "object",
        "required": ["items", "logic_operator"],
        "title": "Filter[SearchColumns]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IDListOperator": {
        "type": "string",
        "enum": ["ID_LIST_CONTAINS", "ID_LIST_NOT_CONTAINS"],
        "title": "IDListOperator"
      },
      "IDOperator": { "type": "string", "enum": ["ID_EQUALS", "ID_NOT_EQUALS"], "title": "IDOperator" },
      "KeywordStat": {
        "properties": {
          "keyword": { "type": "string", "title": "Keyword", "description": "The counted keyword." },
          "filtered_count": {
            "type": "integer",
            "title": "Filtered Count",
            "description": "Number of occurrences of the keyword in the filtered collection"
          },
          "global_count": {
            "type": "integer",
            "title": "Global Count",
            "description": "Number of occurrences of the keyword in the entire collection"
          }
        },
        "type": "object",
        "required": ["keyword", "filtered_count", "global_count"],
        "title": "KeywordStat"
      },
      "ListOperator": { "type": "string", "enum": ["LIST_CONTAINS", "LIST_NOT_CONTAINS"], "title": "ListOperator" },
      "LogicalOperator": {
        "type": "string",
        "enum": ["or", "and"],
        "title": "LogicalOperator",
        "description": "This tells our filter how to combine multiple column expressions."
      },
      "MemoContentQuery": {
        "properties": {
          "proj_id": {
            "type": "integer",
            "title": "Proj Id",
            "description": "The ID of the Project the Memo have to belong to."
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "The ID of the User the Memo have to belong to."
          },
          "starred": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Starred",
            "description": "If set (i.e. not NULL / NONE), only returns Memo that have the given starred status"
          },
          "content_query": {
            "type": "string",
            "minLength": 1,
            "title": "Content Query",
            "description": "The query term to search within the content of the Memo"
          }
        },
        "type": "object",
        "required": ["proj_id", "user_id", "content_query"],
        "title": "MemoContentQuery"
      },
...
      "MemoRead": {
        "properties": {
          "title": { "type": "string", "title": "Title", "description": "Title of the Memo" },
          "content": { "type": "string", "title": "Content", "description": "Content of the Memo" },
          "id": { "type": "integer", "title": "Id", "description": "ID of the Memo" },
          "starred": { "type": "boolean", "title": "Starred", "description": "Starred flag of the Memo" },
          "user_id": { "type": "integer", "title": "User Id", "description": "User the Memo belongs to" },
          "project_id": { "type": "integer", "title": "Project Id", "description": "Project the Memo belongs to" },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created",
            "description": "Created timestamp of the Memo"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "title": "Updated",
            "description": "Updated timestamp of the Memo"
          },
          "attached_object_id": {
            "type": "integer",
            "title": "Attached Object Id",
            "description": "ID of the Object the Memo is attached to"
          },
          "attached_object_type": {
            "allOf": [{ "$ref": "#/components/schemas/AttachedObjectType" }],
            "description": "Type of the Object the ID refers to"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "id",
          "starred",
          "user_id",
          "project_id",
          "created",
          "updated",
          "attached_object_id",
          "attached_object_type"
        ],
        "title": "MemoRead"
      },
...
      "NumberOperator": {
        "type": "string",
        "enum": ["NUMBER_EQUALS", "NUMBER_NOT_EQUALS", "NUMBER_GT", "NUMBER_LT", "NUMBER_GTE", "NUMBER_LTE"],
        "title": "NumberOperator"
      },
      "PaginatedMemoSearchResults": {
        "properties": {
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Flag that indicates whether there are more search results."
          },
          "total": { "type": "integer", "title": "Total", "description": "The total number of results." },
          "current_page_offset": {
            "type": "integer",
            "title": "Current Page Offset",
            "description": "The offset that returns the current results."
          },
          "next_page_offset": {
            "type": "integer",
            "title": "Next Page Offset",
            "description": "The offset that returns the next results."
          },
          "memos": {
            "items": { "$ref": "#/components/schemas/MemoRead" },
            "type": "array",
            "title": "Memos",
            "description": "The Memo search results on the requested page."
          }
        },
        "type": "object",
        "required": ["has_more", "total", "current_page_offset", "next_page_offset", "memos"],
        "title": "PaginatedMemoSearchResults"
      },
...
      "SearchColumns": {
        "type": "string",
        "enum": [
          "SC_SOURCE_DOCUMENT_TYPE",
          "SC_SOURCE_DOCUMENT_FILENAME",
          "SC_DOCUMENT_TAG_ID_LIST",
          "SC_CODE_ID_LIST",
          "SC_USER_ID_LIST",
          "SC_SPAN_ANNOTATIONS"
        ],
        "title": "SearchColumns"
      },
      "SimSearchImageHit": {
        "properties": {
          "sdoc_id": {
            "type": "integer",
            "title": "Sdoc Id",
            "description": "The ID of the SourceDocument similar to the query."
          },
          "score": { "type": "number", "title": "Score", "description": "The similarity score." }
        },
        "type": "object",
        "required": ["sdoc_id", "score"],
        "title": "SimSearchImageHit"
      },
      "SimSearchQuery": {
        "properties": {
          "proj_id": {
            "type": "integer",
            "title": "Proj Id",
            "description": "The ID of the Project the SourceDocuments have to belong to."
          },
          "query": {
            "anyOf": [{ "type": "string" }, { "items": { "type": "string" }, "type": "array" }, { "type": "integer" }],
            "title": "Query",
            "description": "The query term. This can be either a single string, a list of strings for which the average embedding gets computed, or an integer which is interpreted as the SDoc ID of an Image."
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "description": "The number of results to return.",
            "default": 10
          },
          "threshold": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Threshold",
            "description": "The minimum distance to use for the sim search.",
            "default": 0
          }
        },
        "type": "object",
        "required": ["proj_id", "query"],
        "title": "SimSearchQuery"
      },
      "SimSearchSentenceHit": {
        "properties": {
          "sdoc_id": {
            "type": "integer",
            "title": "Sdoc Id",
            "description": "The ID of the SourceDocument similar to the query."
          },
          "score": { "type": "number", "title": "Score", "description": "The similarity score." },
          "sentence_id": {
            "type": "integer",
            "title": "Sentence Id",
            "description": "The sentence id with respect to the SourceDocument"
          }
        },
        "type": "object",
        "required": ["sdoc_id", "score", "sentence_id"],
        "title": "SimSearchSentenceHit"
      },
...
      "SortDirection": { "type": "string", "enum": ["asc", "desc"], "title": "SortDirection" },
...
      "Sort_SearchColumns_": {
        "properties": {
          "column": {
            "anyOf": [{ "$ref": "#/components/schemas/SearchColumns" }, { "type": "integer" }],
            "title": "Column"
          },
          "direction": { "$ref": "#/components/schemas/SortDirection" }
        },
        "type": "object",
        "required": ["column", "direction"],
        "title": "Sort[SearchColumns]"
      },
...
      "SpanEntityStat": {
        "properties": {
          "code_id": {
            "type": "integer",
            "title": "Code Id",
            "description": "The ID of the Code related to the SpanAnnotation"
          },
          "span_text": {
            "type": "string",
            "title": "Span Text",
            "description": "The SpanText the SpanAnnotation spans"
          },
          "filtered_count": {
            "type": "integer",
            "title": "Filtered Count",
            "description": "Number of occurrences of the SpanEntity in a collection of SourceDocuments."
          },
          "global_count": {
            "type": "integer",
            "title": "Global Count",
            "description": "Number of occurrences of the SpanEntity in a collection of SourceDocuments."
          }
        },
        "type": "object",
        "required": ["code_id", "span_text", "filtered_count", "global_count"],
        "title": "SpanEntityStat"
      },
...
      "StringOperator": {
        "type": "string",
        "enum": ["STRING_CONTAINS", "STRING_EQUALS", "STRING_NOT_EQUALS", "STRING_STARTS_WITH", "STRING_ENDS_WITH"],
        "title": "StringOperator"
      },
...
      "TagStat": {
        "properties": {
          "tag": {
            "allOf": [{ "$ref": "#/components/schemas/DocumentTagRead" }],
            "description": "The counted document tag."
          },
          "filtered_count": {
            "type": "integer",
            "title": "Filtered Count",
            "description": "Number of occurrences of the document tag in the filtered documents"
          },
          "global_count": {
            "type": "integer",
            "title": "Global Count",
            "description": "Number of occurrences of the document tag in all documents"
          }
        },
        "type": "object",
        "required": ["tag", "filtered_count", "global_count"],
        "title": "TagStat"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/authentication/login" } }
      }
    }
  }
}